+++ v2/cmd/wails/internal/commands/dev/stdout_scanner.go
@@ -2,11 +2,11 @@ package dev
 
 import (
 	"bufio"
-	"bytes"
-	"github.com/acarl005/stripansi"
 	"net/url"
 	"os"
 	"strings"
+
+	"github.com/acarl005/stripansi"
 )
 
 // stdoutScanner acts as a stdout target that will scan the incoming
@@ -24,17 +24,17 @@ func NewStdoutScanner() *stdoutScanner {
 
 // Write bytes to the scanner. Will copy the bytes to stdout
 func (s *stdoutScanner) Write(data []byte) (n int, err error) {
-	match := bytes.Index(data, []byte("> Local:"))
+	input := stripansi.Strip(string(data))
+	match := strings.Index(input, "Local:")
 	if match != -1 {
-		sc := bufio.NewScanner(strings.NewReader(string(data)))
+		sc := bufio.NewScanner(strings.NewReader(input))
 		for sc.Scan() {
 			line := sc.Text()
 			index := strings.Index(line, "Local:")
 			if index == -1 || len(line) < 7 {
 				continue
 			}
-			line = strings.TrimSpace(line[index+6:])
-			viteServerURL := stripansi.Strip(line)
+			viteServerURL := strings.TrimSpace(line[index+6:])
 			LogGreen("Vite Server URL: %s", viteServerURL)
 			_, err := url.Parse(viteServerURL)
 			if err != nil {