+++ v2/cmd/wails/internal/commands/initialise/templates/templates/svelte-ts/frontend/wailsjs/runtime/runtime.js
@@ -149,6 +149,10 @@ export function WindowSetBackgroundColour(R, G, B, A) {
     window.runtime.WindowSetBackgroundColour(R, G, B, A);
 }
 
+export function ScreenGetAll() {
+    return window.runtime.ScreenGetAll();
+}
+
 export function BrowserOpenURL(url) {
     window.runtime.BrowserOpenURL(url);
 }
@@ -159,8 +163,4 @@ export function Environment() {
 
 export function Quit() {
     window.runtime.Quit();
-}
-
-export function ScreenGetAll() {
-    return window.runtime.ScreenGetAll();
 }