+++ v2/internal/frontend/desktop/darwin/Application.h
@@ -45,7 +45,7 @@ void Quit(void*);
 const char* GetSize(void *ctx);
 const char* GetPosition(void *ctx);
 
-void ProcessURLResponse(void *inctx, const char *url, int statusCode, void *headersString, int headersStringLength, void* data, int datalength);
+void ProcessURLResponse(void *inctx, unsigned long long requestId, int statusCode, void *headersString, int headersStringLength, void* data, int datalength);
 
 /* Dialogs */