+++ v2/internal/github/github.go
@@ -3,11 +3,51 @@ package github
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/charmbracelet/glamour"
 	"io"
 	"net/http"
+	"net/url"
+	"runtime"
 	"sort"
 	"strings"
 )
+
+func GetReleaseNotes(tagVersion string) string {
+	resp, err := http.Get("https://api.github.com/repos/wailsapp/wails/releases/tags/" + url.PathEscape(tagVersion))
+	if err != nil {
+		return "Unable to retrieve release notes. Please check your network connection"
+	}
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return "Unable to retrieve release notes. Please check your network connection"
+	}
+
+	data := map[string]interface{}{}
+	err = json.Unmarshal(body, &data)
+	if err != nil {
+		return "Unable to retrieve release notes. Please check your network connection"
+	}
+
+	if data["body"] == nil {
+		return "No release notes found"
+	}
+
+	result := "# Release Notes for " + tagVersion + "\n" + data["body"].(string)
+	var renderer *glamour.TermRenderer
+	if runtime.GOOS == "windows" {
+		renderer, err = glamour.NewTermRenderer(glamour.WithStyles(glamour.NoTTYStyleConfig))
+	} else {
+		renderer, err = glamour.NewTermRenderer(glamour.WithAutoStyle())
+	}
+	if err != nil {
+		return result
+	}
+	result, err = renderer.Render(result)
+	if err != nil {
+		return err.Error()
+	}
+	return result
+}
 
 // GetVersionTags gets the list of tags on the Wails repo
 // It returns a list of sorted tags in descending order