+++ v2/internal/project/project.go
@@ -22,6 +22,8 @@ type Project struct {
 
 	// Commands used in `wails dev`
 	DevCommand        string `json:"frontend:dev"`
+	DevBuildCommand   string `json:"frontend:dev:build"`
+	DevInstallCommand string `json:"frontend:dev:install"`
 	DevWatcherCommand string `json:"frontend:dev:watcher"`
 	// The url of the external wails dev server. If this is set, this server is used for the frontend. Default ""
 	FrontendDevServerURL string `json:"frontend:dev:serverUrl"`
@@ -84,11 +86,25 @@ type Project struct {
 	NSISType string `json:"nsisType"`
 }
 
-func (p *Project) GetDevInstallerCommand() string {
+func (p *Project) GetDevBuildCommand() string {
+	if p.DevBuildCommand != "" {
+		return p.DevBuildCommand
+	}
 	if p.DevCommand != "" {
 		return p.DevCommand
 	}
+	return p.BuildCommand
+}
+
+func (p *Project) GetDevInstallerCommand() string {
+	if p.DevInstallCommand != "" {
+		return p.DevInstallCommand
+	}
 	return p.InstallCommand
+}
+
+func (p *Project) IsFrontendDevServerURLAutoDiscovery() bool {
+	return p.FrontendDevServerURL == "auto"
 }
 
 func (p *Project) Save() error {