+++ website/src/pages/changelog.md
@@ -7,6 +7,88 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.0.0-beta.43] - 2022-08-08
+
+## Added
+
+* Add support for retrieving the release notes from Github by @leaanthony in https://github.com/wailsapp/wails/pull/1679
+* Add `frontend:dev:install` configuration by @LGiki in https://github.com/wailsapp/wails/pull/1666
+
+## Fixed
+
+* Fix formatting of some error messages by @stffabi in https://github.com/wailsapp/wails/pull/1665
+* Windows dialogs now work when window is not visible yet by @leaanthony in https://github.com/wailsapp/wails/pull/1662
+* Multiple fixes for MacOS asset requests by @stffabi in https://github.com/wailsapp/wails/pull/1668
+  and https://github.com/wailsapp/wails/pull/1681
+* Fix for Go 1.19 by @stffabi in https://github.com/wailsapp/wails/pull/1689
+* Removed Linux warnings by @leaanthony in https://github.com/wailsapp/wails/pull/1656
+* Better support for doubleclick events in drag regions by @leaanthony in https://github.com/wailsapp/wails/pull/1704
+* Allow MacOS frameless window to be miniturisable by @leaanthony in https://github.com/wailsapp/wails/pull/1705
+
+## What's Changed
+
+* add wails-sveltekit-template by @h8gi in https://github.com/wailsapp/wails/pull/1671
+* wails doctor now reports correct MacOS os id by @stffabi in https://github.com/wailsapp/wails/pull/1673
+* Update application-development.mdx by @SamHennessy in https://github.com/wailsapp/wails/pull/1682
+* Move SetMin/Max calls to main thread by @leaanthony in https://github.com/wailsapp/wails/pull/1684
+* Change `frontend:dev` to `frontend:dev:build` by @LGiki in https://github.com/wailsapp/wails/pull/1691
+* Build frontend only before starting the dev watcher command by @stffabi in https://github.com/wailsapp/wails/pull/1694
+* Improve error message for auto dev server discovery without a dev watcher by @stffabi
+  in https://github.com/wailsapp/wails/pull/1711
+
+## New Contributors
+
+* @h8gi made their first contribution in https://github.com/wailsapp/wails/pull/1671
+* @SamHennessy made their first contribution in https://github.com/wailsapp/wails/pull/1682
+
+## [v2.0.0-beta.42] - 2022-07-25
+
+## Added
+
+* Added `options.NewRGBA` and `options.NewRGB` functions to create `*options.RGBA` by @leaanthony
+
+## Fixed
+
+* Fixed initial build of frontend when using `wails dev` on new projects by @leaanthony
+  in https://github.com/wailsapp/wails/pull/1650
+* Ignore empty install command when running `wails dev` by @stffabi in https://github.com/wailsapp/wails/pull/1651
+* Fixed error reporting in templates
+* BackgroundColour documentation fix
+* Generalize manual compile steps [Documentation] by @acheong08 in https://github.com/wailsapp/wails/pull/1644
+
+## [v2.0.0-beta.40] - 2022-07-24
+
+## Added
+
+* Add Show() and Hide() to runtime to show/hide application by @leaanthony
+  in https://github.com/wailsapp/wails/pull/1599
+* Override target platform/arch using GOOS and GOARCH environment variables by @leaanthony
+  in https://github.com/wailsapp/wails/pull/1618
+* Add option to skip frontend rebuild in dev mode by @leaanthony in https://github.com/wailsapp/wails/pull/1632
+
+## Fixed
+
+* Update svelte templates to use vite 3 by @leaanthony in https://github.com/wailsapp/wails/pull/1643
+* Fix plain template by @stffabi in https://github.com/wailsapp/wails/pull/1609
+* Fix Website layout by @leaanthony in https://github.com/wailsapp/wails/pull/1616
+* Fixed typo in documentation page docs/howdoesitwork by @MyNameIsAres in https://github.com/wailsapp/wails/pull/1636
+* Use scaling when setting min/max window by @leaanthony in https://github.com/wailsapp/wails/pull/1557
+
+## Changed
+
+* Install dev dependencies before starting dev mode by @leaanthony in https://github.com/wailsapp/wails/pull/1615
+* Translate and fix website text by @misitebao in https://github.com/wailsapp/wails/pull/1525
+* docs: add MyNameIsAres as a contributor for doc by @allcontributors in https://github.com/wailsapp/wails/pull/1638
+* Deprecate Fullscreen appoption by @acheong08 in https://github.com/wailsapp/wails/pull/1640
+
+## Deprecated
+
+* The `Fullscreen` application option is deprecated. Please
+  use [`WindowStartState`](https://wails.io/docs/reference/options#windowstartstate) instead.
+## New Contributors
+
+* @MyNameIsAres made their first contribution in https://github.com/wailsapp/wails/pull/1636
+
 ## [v2.0.0-beta.39.2] - 2022-07-20
 
 ## Added